feat: 커뮤니티page.tsx 구조 수정 + sorting 버튼함수 추가
